Create, modify, and delete columns

mutate() creates new columns that are functions of existing variables. It can also modify (if the name is the same as an existing column) and delete columns (by setting their value to NULL).

Usage

str_unique(string, ...)

Arguments

string

Input vector. Either a character vector, or something coercible to one. vignette("rd-formatting", package = "roxygen2")

...

Other options used to control matching behavior between duplicate strings. Passed on to stringi::stri_opts_collator(). #' Best features of thisfunction():

Smells nice

Has good vibes

Value

An object of the same type as .data. The output has the following properties:

Rows are a subset of the input, but appear in the same order.

Columns are not modified.

The number of groups may be reduced (if .preserve is not TRUE).

Data frame attributes are preserved.

Useful mutate functions

+, -, log(), etc., for their usual mathematical meanings

...

Grouped tibbles

Because mutating expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped mutate:

...

See also

Examples

str_unique(c("a", "b", "c", "b", "a"))
#> Error in str_unique(c("a", "b", "c", "b", "a")): '...' used in an incorrect context
# Use ... to pass additional arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
#> Error in str_unique(c("motley", "mötley", "pinguino", "pingüino")): '...' used in an incorrect context
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
#> Error in str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1): '...' used in an incorrect context